requirements.txt
python-jenkins
pycrypto
xmltodict
astor
pika
flask
pymongo
pytest==3.8.0
pytest-timeout==1.3.3
